+++ muk_website_llms_txt/doc/changelog.rst
@@ -1,3 +1,9 @@
+`1.2.0`
+-------
+
+- Structured Markdown
+- Response Caching
+
 `1.1.0`
 -------